chore(marketing): webp screenshots, prerender download links, sitemap + og image
- Convert marketing screenshots from PNG to WebP (2.6MB -> 668K) and update refs
- Prerender /download release asset URLs at build time instead of client fetch
- Add og:image and twitter:image meta to /download
- Add /download to sitemap.xml
- Remove dev-only screenshot capture harness
