+++ src/routes/sitemap.xml/+server.ts
@@ -26,6 +26,7 @@ export const GET: RequestHandler = () => {
 
 	const staticPages: SitemapEntry[] = [
 		{ url: SITE_URL, priority: '1.0' },
+		{ url: `${SITE_URL}/download`, priority: '0.8' },
 		{ url: `${SITE_URL}/docs/overview/introduction`, priority: '0.8' },
 		{ url: `${SITE_URL}/blog`, priority: '0.8' }
 	];
